+++ src/App.tsx
@@ -1,25 +1,25 @@
-import { useEffect, useState } from 'react';
 import { Route, Routes } from 'react-router-dom';
-import { useCustomDispatch, useCustomSelector } from './hooks/redax-hooks';
 import styles from './App.module.css';
-import { getRefreshToken } from './api/api';
-import { actions } from './redux/actions';
-import { StoreType } from './store/store';
+import { useCustomDispatch, useCustomSelector } from './hooks/redax-hooks';
 
-import { Greeting } from './components/Greeting/Greeting';
-import NavMenu from './components/NavMenu/NavMenu';
-import Header from './components/Header/Header';
-import NotFound from './components/NotFound/NotFound';
-import Footer from './components/Footer/Footer';
-import Textbook from './components/Textbook/Textbook';
+import { useEffect, useState } from 'react';
+import { getNewTokens } from './api/authApi';
+import { actions } from './redux/actions';
+
 import About from './components/About/About';
+import AudioChallenge from './components/AudioChallenge/AudioChallenge';
+import Authorization from './components/Autorization/Authorization';
+import Footer from './components/Footer/Footer';
+import Greeting from './components/Greeting/Greeting';
+import Header from './components/Header/Header';
+import NavMenu from './components/NavMenu/NavMenu';
+import NotFound from './components/NotFound/NotFound';
 import Sprint from './components/Sprint/Sprint';
 import Statistics from './components/Statistics/Statistics';
-import Authorization from './components/Autorization/Authorization';
-import AudioChallenge from './components/AudioChallenge/AudioChallenge';
+import Textbook from './components/Textbook/Textbook';
 
 function App() {
-  const isNavMenuOpen = useCustomSelector((state: StoreType): boolean => state.navMenu.isNavMenuOpen);
+  const isNavMenuOpen = useCustomSelector((state): boolean => state.navMenu.isNavMenuOpen);
   const [isCheckLogin, setIsCheckLogin] = useState<boolean>(true);
   const token = JSON.parse((localStorage.getItem('token') as string));
   const timeLogin = Number(localStorage.getItem('timeLogin'));
@@ -32,19 +32,17 @@ function App() {
       if (!token) return;
       else if ((Date.now() - timeLogin) < limitTime) dispatch(actions.switchIsLogin(true));
       else {
-        (function refresh() {
-          getRefreshToken().then((response) => {
-            if (response.status === 200) {
-              dispatch(actions.switchIsLogin(true));
-              return response.json().then((res) => {
-                localStorage.setItem('userId', JSON.stringify(res.userId));
-                localStorage.setItem('token', JSON.stringify(res.token));
-                localStorage.setItem('refreshToken', JSON.stringify(res.refreshToken));
-                localStorage.setItem('timeLogin', JSON.stringify(Date.now()));
-              });
-            }
-          });
-        })();
+        getNewTokens().then((response) => {
+          if (response.status === 200) {
+            dispatch(actions.switchIsLogin(true));
+            return response.json().then((res) => {
+              localStorage.setItem('userId', JSON.stringify(res.userId));
+              localStorage.setItem('token', JSON.stringify(res.token));
+              localStorage.setItem('refreshToken', JSON.stringify(res.refreshToken));
+              localStorage.setItem('timeLogin', JSON.stringify(Date.now()));
+            });
+          }
+        });
       } 
     }
   // eslint-disable-next-line react-hooks/exhaustive-deps
